+++ src/App.tsx
@@ -4,7 +4,7 @@ import { Dispatch, Action } from 'redux';
 import { RootState } from './reducers';
 import { webSocketActionCreator } from './actions/WebSocketAction';
 import Watch from './containers/sample/Watch';
-import Title from './calendars/monthly/default/components/Title';
+import Title from './calendars/monthly/default/containers/Title';
 
 interface IStateToProps {
     connected: boolean;
