+++ src/reducers/CurrentReducer.tsx
@@ -5,15 +5,11 @@ import { CurrentAction, CurrentType } from '../actions/CurrentAction';
 export interface ICurrentState {
     nowDate: string,
     nowTime: string,
-    nowMonth: string,
-    nowMonthName: string,
 }
 
 const initState: ICurrentState = {
     nowDate: '2019-05-01',
     nowTime: '12:00:00',
-    nowMonth: '5',
-    nowMonthName: 'May',
 };
 
 export const CurrentReducer: Reducer<ICurrentState, CurrentAction> = (
@@ -26,8 +22,6 @@ export const CurrentReducer: Reducer<ICurrentState, CurrentAction> = (
             return {
                 nowDate: nowMoment.format('YYYY-MM-DD'),
                 nowTime: nowMoment.format('hh:mm:ss'),
-                nowMonth: nowMoment.format('M'),
-                nowMonthName: nowMoment.format('MMMM'),
             };
         }
         default: {
